# SIMD: replace raw SIMD of ceil with universal intrinsics